Update with request changes
average now is in bigint

+++ mosaicod/crates/mosaicod-core/src/types/resources.rs
@@ -279,7 +279,7 @@ pub struct TopicChunksStats {
     pub total_size_bytes: i64,
     pub total_row_count: i64,
     /// Average row footprint in uncompressed Arrow bytes
-    pub avg_bytes_per_row: f64,
+    pub avg_bytes_per_row: i64,
 }
 
 /// Metadata properties associated to a topic.
